Update django provider dianostics queue locations
