+++ server/src/providers/django/djangoProjectDetector.ts
