+++ server/src/services/diagnostics/index.ts
@@ -1,1 +1,2 @@
 export * from "./diagnosticsManager";
+export * from "./diagnosticsQueue";
